# Speed up tests on K8s by not blocking until test namespace is gone