**What type of PR is this:**
/kind tests
/area testing

**What does this PR do / why we need it:**
When running against my local K8s cluster, I noticed that we were spending several seconds just waiting for the test namespace to be gone, as part of the tear-down logic.
However, when running against an OpenShift cluster, project/namespace deletion appears not to be blocking.
I might be missing something, but I was wondering if it wouldn't make sense to have the same behavior in `helper_kubectl.go`.

This might not have any notable impact in CI, but I noticed substantial gains when iterating locally.

Example with a simple test suite (execution time went from ~116 seconds to ~25 seconds with the test namespace deleted asynchronously):

**Before these changes**
```
❯ time ginkgo -focus 'odo devfile init command tests' tests/integration/devfile                                                                                                                                                                                                  
Running Suite: Devfile Suite                                                                                                                                                                                                                                                     
============================                                                                                                                                                                                                                                                     
Random Seed: 1653660542                                                                                                                                                                                                                                                          
Will run 18 of 147 specs   

•••••••••••••••••                                                   

Ran 18 of 147 Specs in 116.800 seconds
SUCCESS! -- 18 Passed | 0 Failed | 0 Pending | 129 Skipped
PASS

Ginkgo ran 1 suite in 1m58.857849553s
Test Suite Passed
ginkgo -focus 'odo devfile init command tests' tests/integration/devfile  11.99s user 3.18s system 12% cpu 1:58.86 total
```

**With these changes**
```
❯ time ginkgo -focus 'odo devfile init command tests' tests/integration/devfile 
Running Suite: Devfile Suite
============================
Random Seed: 1653660125
Will run 18 of 147 specs

•••••••••••••••••

Ran 18 of 147 Specs in 24.615 seconds
SUCCESS! -- 18 Passed | 0 Failed | 0 Pending | 129 Skipped
PASS

Ginkgo ran 1 suite in 26.710032151s
Test Suite Passed
ginkgo -focus 'odo devfile init command tests' tests/integration/devfile  11.21s user 2.76s system 52% cpu 26.713 total
```


**Which issue(s) this PR fixes:**
No related issue

**PR acceptance criteria:**

- [ ] Unit test 

- [ ] Integration test 

- [ ] Documentation 

**How to test changes / Special notes to the reviewer:**